Finish breaking down FactoryProduction into pieces
