+++ src/main/java/org/chainoptim/desktop/shared/util/JsonUtil.java
@@ -1,5 +1,6 @@
 package org.chainoptim.desktop.shared.util;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
@@ -7,6 +8,7 @@ import lombok.Getter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.apache.commons.text.StringEscapeUtils;
 
 /*
  * Util to configure Json (de)serialization
@@ -28,6 +30,17 @@ public class JsonUtil {
 
         // Register implicitly
         objectMapper.findAndRegisterModules();
+    }
+
+    public static <T> String prepareJsonString(T data) {
+        String jsonString = "{}";
+        try {
+            jsonString = JsonUtil.getObjectMapper().writeValueAsString(data);
+        } catch (JsonProcessingException ex) {
+            ex.printStackTrace();
+        }
+        String finalJsonString = jsonString;
+        return StringEscapeUtils.escapeEcmaScript(finalJsonString);
     }
 
 }
