+++ src/main/resources/org/chainoptim/desktop/features/factory/FactoryProductionView.fxml
@@ -1,92 +1,21 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?import javafx.scene.layout.*?>
+
 <?import javafx.scene.control.SplitPane?>
 <?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.collections.FXCollections?>
-<?import java.lang.String?>
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="org.chainoptim.desktop.features.factory.controller.FactoryProductionController">
-
+            fx:controller="org.chainoptim.desktop.features.factory.controller.FactoryProductionController"
+>
     <SplitPane dividerPositions="0.75" prefHeight="9999" VBox.vgrow="ALWAYS" AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
-        <!-- Graph Container Side -->
+        <!-- Tabs Container Side -->
         <ScrollPane fitToWidth="true" fitToHeight="true">
-            <StackPane fx:id="graphContainer"/>
+            <StackPane fx:id="tabsContainer" VBox.vgrow="ALWAYS"/>
         </ScrollPane>
 
-        <!-- Toolbar Side -->
+        <!-- Toolbar Container Side -->
         <ScrollPane fitToWidth="true" fitToHeight="true">
-            <VBox spacing="10" style="-fx-padding: 10; -fx-background-color: #ffffff;">
-                <Label text="Toolbar" style="-fx-font-size: 18px; -fx-font-weight: bold"/>
-
-                <!-- Section: Edit Configuration -->
-                <HBox alignment="CENTER_LEFT" spacing="10">
-                    <Label text="Edit Configuration" style="-fx-font-size: 14px; -fx-font-weight: bold"/>
-                    <Button fx:id="toggleEditConfigurationButton" onAction="#toggleEditConfigurationSection" styleClass="no-style-button"/>
-                </HBox>
-                <VBox spacing="10">
-                    <VBox fx:id="editConfigurationContentVBox" spacing="10">
-                        <Button text="Add Stage"/>
-                        <Button text="Edit Stage"/>
-                        <Button text="Add Connection"/>
-                    </VBox>
-                </VBox>
-
-                <!-- Section: Display info -->
-                <HBox alignment="CENTER_LEFT" spacing="10">
-                    <Label text="Display Info" style="-fx-font-size: 14px; -fx-font-weight: bold"/>
-                    <Button fx:id="toggleDisplayInfoButton" onAction="#toggleDisplayInfoSection" styleClass="no-style-button"/>
-                </HBox>
-                <VBox spacing="10">
-                    <VBox fx:id="displayInfoContentVBox" spacing="10">
-                        <CheckBox fx:id="quantitiesCheckBox" text="Quantities (Q)"/>
-                        <CheckBox fx:id="capacityCheckBox" text="Capacities (C)"/>
-                        <CheckBox fx:id="priorityCheckBox" text="Priorities (P)" />
-                    </VBox>
-                </VBox>
-
-                <!-- Section: Resource Allocation -->
-                <HBox alignment="CENTER_LEFT" spacing="10">
-                    <Label text="Allocate Resources" style="-fx-font-size: 14px; -fx-font-weight: bold"/>
-                    <Button fx:id="toggleResourceAllocationButton" onAction="#toggleResourceAllocationSection" styleClass="no-style-button"/>
-                </HBox>
-                <VBox spacing="10">
-                    <VBox fx:id="resourceAllocationContentBox" spacing="10">
-                        <Label text="From Inventory" style="-fx-font-size: 12px;"/>
-                        <HBox spacing="10">
-                            <TextField fx:id="resourceAllocationInput" promptText="Enter number" style="-fx-pref-width: 100px;" />
-                            <ComboBox fx:id="timePeriodSelect" style="-fx-pref-width: 150px;">
-                                <items>
-                                    <FXCollections fx:factory="observableArrayList">
-                                        <String fx:value="Hours"/>
-                                        <String fx:value="Days"/>
-                                        <String fx:value="Weeks"/>
-                                        <String fx:value="Months"/>
-                                    </FXCollections>
-                                </items>
-                            </ComboBox>
-                        </HBox>
-                        <Button text="Allocate Resources" onAction="#handleAllocateResources"/>
-
-                    </VBox>
-                </VBox>
-
-                <!-- Section: Seek resources -->
-                <HBox alignment="CENTER_LEFT" spacing="10">
-                    <Label text="Seek Resources" style="-fx-font-size: 14px; -fx-font-weight: bold"/>
-                    <Button fx:id="toggleSeekResourcesButton" onAction="#toggleSeekResourcesSection" styleClass="no-style-button"/>
-                </HBox>
-                <VBox spacing="10">
-                    <VBox fx:id="seekResourcesContentBox" spacing="10">
-                        <Button text="Seek Missing Resources"/>
-                    </VBox>
-                </VBox>
-            </VBox>
+            <StackPane fx:id="toolbarContainer" VBox.vgrow="ALWAYS"/>
         </ScrollPane>
     </SplitPane>
 </AnchorPane>
