--- src/main/resources/org/chainoptim/desktop/features/factory/FactoryProductionViewNew.fxml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?import javafx.scene.layout.*?>
-
-<?import javafx.scene.control.SplitPane?>
-<?import javafx.scene.control.ScrollPane?>
-<AnchorPane xmlns="http://javafx.com/javafx"
-            xmlns:fx="http://javafx.com/fxml"
-            fx:controller="org.chainoptim.desktop.features.factory.controller.FactoryProductionControllerNew"
->
-    <SplitPane dividerPositions="0.75" prefHeight="9999" VBox.vgrow="ALWAYS" AnchorPane.topAnchor="0" AnchorPane.bottomAnchor="0" AnchorPane.leftAnchor="0" AnchorPane.rightAnchor="0">
-        <!-- Tabs Container Side -->
-        <ScrollPane fitToWidth="true" fitToHeight="true">
-            <StackPane fx:id="tabsContainer" VBox.vgrow="ALWAYS"/>
-        </ScrollPane>
-
-        <!-- Toolbar Container Side -->
-        <ScrollPane fitToWidth="true" fitToHeight="true">
-            <StackPane fx:id="toolbarContainer" VBox.vgrow="ALWAYS"/>
-        </ScrollPane>
-    </SplitPane>
-</AnchorPane>
