+++ src/main/resources/org/chainoptim/desktop/features/factory/factoryproduction/CreateStageView.fxml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?import javafx.scene.layout.*?>
 
+<?import javafx.scene.control.Label?>
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="org.chainoptim.desktop.features.factory.controller.factoryproduction.CreateStageController"
-            prefHeight="400.0" prefWidth="600.0">
+            fx:controller="org.chainoptim.desktop.features.factory.controller.factoryproduction.CreateStageController">
 
+    <Label text="Create Stage"/>
 </AnchorPane>
